==> 3_playground.ipynb <==
from pyspark.sql import SparkSession

# Create a SparkSession
spark = SparkSession.builder.appName("algo").getOrCreate()

df = spark.read.format("csv").option("header", "true").option("sep", ";").load("out_prod_cartesien.csv")


df.write.mode("overwrite").parquet("out_prod_cartesien.parquet")


df = spark.read.format("parquet").option("header", "true").load("out_prod_cartesien.parquet")


df.count()

df.show(100)

distinct_codes = df.select('CODE_PRODUIT_STOCKAGE').distinct()
distinct_codes.show()


distinct_codes.count()

filtered_df = df.where(df["CODE_SITE_ORIGINE"] == "D681").select('CODE_PRODUIT_STOCKAGE').distinct()
filtered_df.show()

filtered_df
